requirements.txt
numpy
matplotlib
Pillow
tqdm
torchsummary